@@ -18,10 +18,10 @@
     "form-data": "^2.3.1",
     "jquery": "^2.2.1",
     "muicss": "^0.9.27",
-    "react": "^0.14.7",
+    "react": "^16.0.0",
     "react-bootstrap": "^0.31.2",
     "react-bs-notifier": "^4.3.2",
-    "react-dom": "^0.14.7",
+    "react-dom": "^16.0.0",
     "react-redux": "^5.0.6",
     "react-router": "^2.8.1",
     "react-router-dom": "^4.2.2",
